requirements.txt
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-chroma>=0.1.4
langchain-community>=0.3.7
chromadb>=0.5.20
openai>=1.54.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0
tiktoken>=0.8.0
streamlit>=1.40.1
pandas>=2.2.3
tqdm>=4.66.6